Add GTL[(K, V)].toMultiMap[F[_]]: MultiMap[F, K, V]
It's somewhat arbitrary where to define 'toMultiMap', it only depends on 'map'
and a CanBuildFrom instance (that actually builds the multimap), but map
is defined on FilterMonadic, GenTraversableLike, & TraversableLike.

So I've changed the list & set pimps to only extend genTraversableLike (i.e. no longer
extend filterMonadic, because otherwise the pimp is ambiguous), but I'm leaving the
FilterMonadic based definition (now no longer extended by anything but it's own companion
object) for backwards compatibility.